Modifications for Snowstorm production
Most of the modifications are to the convert_edepsim script to generate
varied parameter values for Snowstorm.

Also included a new RNG (SplitMix) for the random numbers for
portability reasons.

For changes to larnd-sim and nd-flow, see their respective repositories.

Minor changes to several run scripts, most importantly disablling the
calib_final_hit workflow in nd-flow as it was causing errors for some
files.